pyproject.toml
[project]
name = "agriculture-cameroun"
version = "0.1.0"
description = "Système Multi-Agents pour l'Agriculture Camerounaise"
authors = [{ name = "Mbassi Loic Aron", email = "wwwmbassiloic@gmail.com" }]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.11.3",
    "aiohttp>=3.12.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "numpy>=2.2.0",
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langfuse>=2.0.0",
    "sse-starlette>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["agriculture_cameroun", "evaluation", "tests"]
exclude = ["poetry.lock", "*.pyi"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line too long — géré par le formatter
    "N806",  # uppercase variable names (constantes locales)
    "B008",  # function-call in default arg (pattern ADK/FastAPI)
]

[tool.ruff.lint.isort]
known-first-party = ["agriculture_cameroun", "evaluation"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
asyncio_mode = "auto"
filterwarnings = ["ignore::UserWarning"]

# ── Build ─────────────────────────────────────────────────────────────────────
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agriculture_cameroun"]
